Cache last background image
makes background images available much faster on slow networks because
it uses the last sessions image that it saved to disk, it does not have
to do the full url grab and parse to get the background art first. This
can take a while on slow networks or slow systems.