Show the corresponding media stubs message
When playing media stubs file, nothing happens. A dialog asking you to insert the disc should open.
There was a typo in the **IsPlaceHolder** property and the wrong URL was returned by the **getPlayUrl()** function.

This fix the **PLAY** method by showing the correct dialog. I didn't do the same with **PLAYPlaylist** because I'm not sure what it does. For the **PLAYPlaylist**, the "This item is not playable" dialog will appear, which it wasn't showing because of the typos.

This came from here: http://mediabrowser.tv/community/index.php?/topic/10103-show-media-stubs-message/