Refined the new token strategy
If the user is changed within the add-on configs, a new token is automatically generated. A restart is however required to get the content to reflect the change. However, I've noticed that without the refresh, at episode level it gets the proper user. The problem is that the seasons don't seem to refresh. To review.